Split proxy into multiple header files
