+++ CMakeLists.txt
@@ -24,20 +24,15 @@ if(PROJECT_IS_TOP_LEVEL)
   )
 endif()
 
-target_sources(
-  msft_proxy4
+file(GLOB_RECURSE proxy_public_headers CONFIGURE_DEPENDS
+     "${CMAKE_CURRENT_SOURCE_DIR}/include/proxy/*.h"
+     "${CMAKE_CURRENT_SOURCE_DIR}/include/proxy/*.ixx")
+target_sources(msft_proxy4
   INTERFACE
     FILE_SET public_headers
-      TYPE HEADERS
-      BASE_DIRS include
-      FILES
-        include/proxy/proxy.h
-        include/proxy/proxy_macros.h
-        include/proxy/proxy_fmt.h
-        include/proxy/v4/proxy.ixx
-        include/proxy/v4/proxy.h
-        include/proxy/v4/proxy_macros.h
-        include/proxy/v4/proxy_fmt.h
+    TYPE HEADERS
+    BASE_DIRS include
+    FILES ${proxy_public_headers}
 )
 
 target_compile_features(msft_proxy4 INTERFACE cxx_std_20)
