+++ tests/CMakeLists.txt
@@ -14,7 +14,7 @@ include(GoogleTest)
 add_executable(
   msft_proxy_tests
   proxy_creation_tests.cpp
-  proxy_details_tests.cpp
+  proxy_detail_tests.cpp
   proxy_dispatch_tests.cpp
   proxy_fmt_format_no_wchar_tests.cpp
   proxy_fmt_format_tests.cpp
