+++ tests/proxy_rtti_tests.cpp
@@ -6,27 +6,27 @@
 #include <proxy/proxy.h>
 #include <vector>
 
-namespace proxy_rtti_tests_details {
+namespace proxy_rtti_tests_detail {
 
 struct TestFacade : pro::facade_builder                   //
                     ::add_skill<pro::skills::rtti>        //
                     ::add_skill<pro::skills::direct_rtti> //
                     ::build {};
 
-} // namespace proxy_rtti_tests_details
+} // namespace proxy_rtti_tests_detail
 
-namespace details = proxy_rtti_tests_details;
+namespace detail = proxy_rtti_tests_detail;
 
 TEST(ProxyRttiTests, TestIndirectCast_Ref_Succeed) {
   int v = 123;
-  pro::proxy<details::TestFacade> p = &v;
+  pro::proxy<detail::TestFacade> p = &v;
   proxy_cast<int&>(*p) = 456;
   ASSERT_EQ(v, 456);
 }
 
 TEST(ProxyRttiTests, TestIndirectCast_Ref_Fail) {
   int v = 123;
-  pro::proxy<details::TestFacade> p = &v;
+  pro::proxy<detail::TestFacade> p = &v;
   bool exception_thrown = false;
   try {
     proxy_cast<double&>(*p);
@@ -38,7 +38,7 @@ TEST(ProxyRttiTests, TestIndirectCast_Ref_Fail) {
 
 TEST(ProxyRttiTests, TestIndirectCast_ConstRef_Succeed) {
   int v = 123;
-  pro::proxy<details::TestFacade> p = &v;
+  pro::proxy<detail::TestFacade> p = &v;
   const int& r = proxy_cast<const int&>(*p);
   ASSERT_EQ(&v, &r);
   ASSERT_EQ(v, 123);
@@ -46,7 +46,7 @@ TEST(ProxyRttiTests, TestIndirectCast_ConstRef_Succeed) {
 
 TEST(ProxyRttiTests, TestIndirectCast_ConstRef_Fail) {
   int v = 123;
-  pro::proxy<details::TestFacade> p = &v;
+  pro::proxy<detail::TestFacade> p = &v;
   bool exception_thrown = false;
   try {
     proxy_cast<const double&>(*p);
@@ -58,7 +58,7 @@ TEST(ProxyRttiTests, TestIndirectCast_ConstRef_Fail) {
 
 TEST(ProxyRttiTests, TestIndirectCast_Copy_Succeed) {
   int v1 = 123;
-  pro::proxy<details::TestFacade> p = &v1;
+  pro::proxy<detail::TestFacade> p = &v1;
   int v2 = proxy_cast<int>(*p);
   ASSERT_EQ(v1, 123);
   ASSERT_EQ(v2, 123);
@@ -66,7 +66,7 @@ TEST(ProxyRttiTests, TestIndirectCast_Copy_Succeed) {
 
 TEST(ProxyRttiTests, TestIndirectCast_Copy_Fail) {
   int v = 123;
-  pro::proxy<details::TestFacade> p = &v;
+  pro::proxy<detail::TestFacade> p = &v;
   bool exception_thrown = false;
   try {
     proxy_cast<double>(*p);
@@ -78,14 +78,14 @@ TEST(ProxyRttiTests, TestIndirectCast_Copy_Fail) {
 
 TEST(ProxyRttiTests, TestIndirectCast_Move_Succeed) {
   const std::vector<int> v1{1, 2, 3};
-  auto p = pro::make_proxy<details::TestFacade>(v1);
+  auto p = pro::make_proxy<detail::TestFacade>(v1);
   auto v2 = proxy_cast<std::vector<int>>(std::move(*p));
   ASSERT_EQ(v2, v1);
   ASSERT_FALSE(p.has_value());
 }
 
 TEST(ProxyRttiTests, TestIndirectCast_Move_Fail) {
-  auto p = pro::make_proxy<details::TestFacade>(std::vector<int>{1, 2, 3});
+  auto p = pro::make_proxy<detail::TestFacade>(std::vector<int>{1, 2, 3});
   bool exception_thrown = false;
   try {
     proxy_cast<std::vector<double>>(std::move(*p));
@@ -97,7 +97,7 @@ TEST(ProxyRttiTests, TestIndirectCast_Move_Fail) {
 
 TEST(ProxyRttiTests, TestIndirectCast_Ptr_Succeed) {
   int v = 123;
-  pro::proxy<details::TestFacade> p = &v;
+  pro::proxy<detail::TestFacade> p = &v;
   auto ptr = proxy_cast<int>(&*p);
   static_assert(std::is_same_v<decltype(ptr), int*>);
   ASSERT_EQ(ptr, &v);
@@ -106,7 +106,7 @@ TEST(ProxyRttiTests, TestIndirectCast_Ptr_Succeed) {
 
 TEST(ProxyRttiTests, TestIndirectCast_Ptr_Fail) {
   int v = 123;
-  pro::proxy<details::TestFacade> p = &v;
+  pro::proxy<detail::TestFacade> p = &v;
   auto ptr = proxy_cast<double>(&*p);
   static_assert(std::is_same_v<decltype(ptr), double*>);
   ASSERT_EQ(ptr, nullptr);
@@ -115,7 +115,7 @@ TEST(ProxyRttiTests, TestIndirectCast_Ptr_Fail) {
 
 TEST(ProxyRttiTests, TestIndirectCast_ConstPtr_Succeed) {
   int v = 123;
-  pro::proxy<details::TestFacade> p = &v;
+  pro::proxy<detail::TestFacade> p = &v;
   auto ptr = proxy_cast<const int>(&*p);
   static_assert(std::is_same_v<decltype(ptr), const int*>);
   ASSERT_EQ(ptr, &v);
@@ -124,7 +124,7 @@ TEST(ProxyRttiTests, TestIndirectCast_ConstPtr_Succeed) {
 
 TEST(ProxyRttiTests, TestIndirectCast_ConstPtr_Fail) {
   int v = 123;
-  pro::proxy<details::TestFacade> p = &v;
+  pro::proxy<detail::TestFacade> p = &v;
   auto ptr = proxy_cast<const double>(&*p);
   static_assert(std::is_same_v<decltype(ptr), const double*>);
   ASSERT_EQ(ptr, nullptr);
@@ -133,20 +133,20 @@ TEST(ProxyRttiTests, TestIndirectCast_ConstPtr_Fail) {
 
 TEST(ProxyRttiTests, TestIndirectTypeid) {
   int a = 123;
-  pro::proxy<details::TestFacade> p = &a;
+  pro::proxy<detail::TestFacade> p = &a;
   ASSERT_EQ(proxy_typeid(*p), typeid(int));
 }
 
 TEST(ProxyRttiTests, TestDirectCast_Ref_Succeed) {
   int v = 123;
-  pro::proxy<details::TestFacade> p = &v;
+  pro::proxy<detail::TestFacade> p = &v;
   *proxy_cast<int*&>(p) = 456;
   ASSERT_EQ(v, 456);
 }
 
 TEST(ProxyRttiTests, TestDirectCast_Ref_Fail) {
   int v = 123;
-  pro::proxy<details::TestFacade> p = &v;
+  pro::proxy<detail::TestFacade> p = &v;
   bool exception_thrown = false;
   try {
     proxy_cast<double*&>(p);
@@ -158,7 +158,7 @@ TEST(ProxyRttiTests, TestDirectCast_Ref_Fail) {
 
 TEST(ProxyRttiTests, TestDirectCast_ConstRef_Succeed) {
   int v = 123;
-  pro::proxy<details::TestFacade> p = &v;
+  pro::proxy<detail::TestFacade> p = &v;
   int* const& r = proxy_cast<int* const&>(p);
   ASSERT_EQ(&v, r);
   ASSERT_EQ(v, 123);
@@ -166,7 +166,7 @@ TEST(ProxyRttiTests, TestDirectCast_ConstRef_Succeed) {
 
 TEST(ProxyRttiTests, TestDirectCast_ConstRef_Fail) {
   int v = 123;
-  pro::proxy<details::TestFacade> p = &v;
+  pro::proxy<detail::TestFacade> p = &v;
   bool exception_thrown = false;
   try {
     proxy_cast<double* const&>(p);
@@ -178,7 +178,7 @@ TEST(ProxyRttiTests, TestDirectCast_ConstRef_Fail) {
 
 TEST(ProxyRttiTests, TestDirectCast_Copy_Succeed) {
   int v1 = 123;
-  pro::proxy<details::TestFacade> p = &v1;
+  pro::proxy<detail::TestFacade> p = &v1;
   int* v2 = proxy_cast<int*>(p);
   ASSERT_EQ(v2, &v1);
   ASSERT_EQ(v1, 123);
@@ -186,7 +186,7 @@ TEST(ProxyRttiTests, TestDirectCast_Copy_Succeed) {
 
 TEST(ProxyRttiTests, TestDirectCast_Copy_Fail) {
   int v = 123;
-  pro::proxy<details::TestFacade> p = &v;
+  pro::proxy<detail::TestFacade> p = &v;
   bool exception_thrown = false;
   try {
     proxy_cast<double*>(p);
@@ -198,7 +198,7 @@ TEST(ProxyRttiTests, TestDirectCast_Copy_Fail) {
 
 TEST(ProxyRttiTests, TestDirectCast_Move_Succeed) {
   int v1 = 123;
-  pro::proxy<details::TestFacade> p = &v1;
+  pro::proxy<detail::TestFacade> p = &v1;
   auto v2 = proxy_cast<int*>(std::move(p));
   ASSERT_EQ(v2, &v1);
   ASSERT_EQ(v1, 123);
@@ -207,7 +207,7 @@ TEST(ProxyRttiTests, TestDirectCast_Move_Succeed) {
 
 TEST(ProxyRttiTests, TestDirectCast_Move_Fail) {
   int v1 = 123;
-  pro::proxy<details::TestFacade> p = &v1;
+  pro::proxy<detail::TestFacade> p = &v1;
   bool exception_thrown = false;
   try {
     proxy_cast<double*>(std::move(p));
@@ -220,7 +220,7 @@ TEST(ProxyRttiTests, TestDirectCast_Move_Fail) {
 
 TEST(ProxyRttiTests, TestDirectCast_Ptr_Succeed) {
   int v = 123;
-  pro::proxy<details::TestFacade> p = &v;
+  pro::proxy<detail::TestFacade> p = &v;
   auto ptr = proxy_cast<int*>(&p);
   static_assert(std::is_same_v<decltype(ptr), int**>);
   ASSERT_EQ(*ptr, &v);
@@ -229,7 +229,7 @@ TEST(ProxyRttiTests, TestDirectCast_Ptr_Succeed) {
 
 TEST(ProxyRttiTests, TestDirectCast_Ptr_Fail) {
   int v = 123;
-  pro::proxy<details::TestFacade> p = &v;
+  pro::proxy<detail::TestFacade> p = &v;
   auto ptr = proxy_cast<double*>(&p);
   static_assert(std::is_same_v<decltype(ptr), double**>);
   ASSERT_EQ(ptr, nullptr);
@@ -238,7 +238,7 @@ TEST(ProxyRttiTests, TestDirectCast_Ptr_Fail) {
 
 TEST(ProxyRttiTests, TestDirectCast_ConstPtr_Succeed) {
   int v = 123;
-  pro::proxy<details::TestFacade> p = &v;
+  pro::proxy<detail::TestFacade> p = &v;
   auto ptr = proxy_cast<int* const>(&p);
   static_assert(std::is_same_v<decltype(ptr), int* const*>);
   ASSERT_EQ(*ptr, &v);
@@ -247,7 +247,7 @@ TEST(ProxyRttiTests, TestDirectCast_ConstPtr_Succeed) {
 
 TEST(ProxyRttiTests, TestDirectCast_ConstPtr_Fail) {
   int v = 123;
-  pro::proxy<details::TestFacade> p = &v;
+  pro::proxy<detail::TestFacade> p = &v;
   auto ptr = proxy_cast<double* const>(&*p);
   static_assert(std::is_same_v<decltype(ptr), double* const*>);
   ASSERT_EQ(ptr, nullptr);
@@ -256,6 +256,6 @@ TEST(ProxyRttiTests, TestDirectCast_ConstPtr_Fail) {
 
 TEST(ProxyRttiTests, TestDirectTypeid) {
   int a = 123;
-  pro::proxy<details::TestFacade> p = &a;
+  pro::proxy<detail::TestFacade> p = &a;
   ASSERT_EQ(proxy_typeid(p), typeid(int*));
 }
